Minor bug fix, Ability to use new Plaques, Added Icon
Fixed bug that caused "Settings" window not to come up sometimes.  Added the ability to make as many of your Steam Games into the new Plaque shortcuts in VRToolbox.  Added an Icon to the exe: thank you, z0oinks!
Note that this file is NOT included in the full zip file.  Use this to replace the exe that you already have.